Overhaul repo for gnome 45.
There's been a pile of changes done from the chroot testing grounds that
are now too many for me to care to commit them all separately. Here is a
mass repo edit to push this to gnome 45.

Note: Several packages are still from the gnome 44 tree, due to the
xdg-desktop-portal version in slackware-current.