Several fixes for the December 2023 update.
xvfb-run source was 404, so update that to one that is available and
matching slackware-current's version.

gnome-all was updated to use python3-toml instead of python-toml.

gupnp-av at the latest version requires a patch to build with the latest
libxml versions.

libshumate required an updated md5sum.

webkit2gtk is updated to match the other versions on SBo.